refactor ViewsHelperFunctionsTestCase to `test_models.KippProjectMethodsTestCase` (since function was migrated to a KippoProject method)
